feat: :sparkles: Move game form to a separate component
